Update common code
